clean, and organized project structure
