+++ src/components/AboutUsPage/AboutHero.css
