+++ src/components/Admin/AdminConsts.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { jwtDecode } from "jwt-decode";
 import { Pencil, Save, Loader2, LogOut, Home, ChevronDown, ChevronUp } from "lucide-react";
-import "../styles/AdminConsts.css";
+import "./AdminConsts.css";
 
 const API_BASE = "https://financesmarttools-backend.onrender.com";
