+++ src/components/Admin/AdminPage.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import "../styles/AdminPage.css";
+import "./AdminPage.css";
 import { useNavigate } from "react-router-dom";
 import { jwtDecode } from "jwt-decode";
 import {
