+++ src/components/Calculators/IncomeTaxWithPoints.jsx
@@ -5,7 +5,7 @@ import {
     ChevronUp,
     AlertTriangle,
 } from "lucide-react";
-import "../styles/IncomeTaxWithPoints.css";
+import "./IncomeTaxWithPoints.css";
 
 const API_BASE = "https://financesmarttools-backend.onrender.com";
