+++ src/components/Calculators/MicroSelfEmployedCalculator.jsx
@@ -8,11 +8,11 @@ import {
     ChevronUp,
     AlertTriangle,
 } from "lucide-react";
-import "../styles/SelfEmployedCost.css";
+import "./MicroSelfEmployedCalculator.css";
 
 const API_BASE = "https://financesmarttools-backend.onrender.com";
 
-export default function SelfEmployedCost() {
+export default function MicroSelfEmployedCalculator() {
     const navigate = useNavigate();
 
     const [yearlyIncome, setYearlyIncome] = useState("");
@@ -32,16 +32,20 @@ export default function SelfEmployedCost() {
             setAuthError(true);
             return;
         }
+
+        // ✅ validate input
         if (!yearlyIncome || isNaN(yearlyIncome) || parseFloat(yearlyIncome) <= 0) {
             alert("אנא הזן הכנסה שנתית תקינה");
             return;
         }
+
         setLoading(true);
         setAuthError(false);
+        setExpanded(false);
 
         try {
             const res = await fetch(
-                `${API_BASE}/self-employed/self-employed?token=${encodeURIComponent(token)}`,
+                `${API_BASE}/micro-self-employed?token=${encodeURIComponent(token)}`,
                 {
                     method: "POST",
                     headers: { "Content-Type": "application/json" },
@@ -66,19 +70,24 @@ export default function SelfEmployedCost() {
         }
     };
 
+    const fmt = (n) =>
+        typeof n === "number"
+            ? n.toLocaleString("he-IL", { maximumFractionDigits: 2 })
+            : n;
+
     return (
         <div className="calcpage" dir="rtl">
 
 
             {/* Intro */}
             <section className="calcpage-intro">
-                <h1>מחשבון עצמאי</h1>
+                <h1>מחשבון עצמאי זעיר</h1>
                 <p className="calcpage-tagline">
-                    רוצים לדעת כמה תשלמו לביטוח לאומי כעצמאים? המחשבון הזה בשבילכם.
+                    כלי מהיר ופשוט לבדיקת תשלומי ביטוח לאומי ובריאות לעצמאי זעיר.
                 </p>
                 <div className="calcpage-hero-box">
-                    כאן תגלו את החבות החודשית והשנתית לביטוח לאומי, בהתאם להכנסה השנתית שלכם.
-                    פשוט הזינו את ההכנסה השנתית ותראו את התוצאה מיד.
+                    המחשבון מחשב באופן מיידי את ההפרשות על בסיס הכנסה חייבת (לפי 70% מהברוטו),
+                    ומציג לכם כמה תשלמו בחודש ובשנה — ומה נשאר נטו לאחר הניכויים.
                 </div>
             </section>
 
@@ -87,14 +96,13 @@ export default function SelfEmployedCost() {
                 <h2>🧮 בצעו חישוב</h2>
                 <div className="calcpage-form-grid">
                     <div className="calcpage-input-group">
-                        <label>הכנסה שנתית</label>
+                        <label>הכנסה שנתית (ברוטו)</label>
                         <input
                             type="number"
                             min="0"
-                            step="1"
                             value={yearlyIncome}
                             onChange={(e) => setYearlyIncome(e.target.value)}
-                            placeholder="הכנס הכנסה שנתית..."
+                            placeholder="לדוגמה: 120000"
                             required
                         />
                     </div>
@@ -103,10 +111,27 @@ export default function SelfEmployedCost() {
                 <button
                     onClick={handleCalculate}
                     className="calcpage-btn submit"
-                    disabled={loading}
+                    disabled={loading || !yearlyIncome}
                 >
                     {loading ? "מחשב..." : "חשב"}
                 </button>
+
+                {/* Footer buttons */}
+                <div className="calcpage-form-footer">
+                    <button onClick={() => navigate(-1)} className="calcpage-btn home">
+                        🔙 חזור
+                    </button>
+                    <button
+                        onClick={() => {
+                            setYearlyIncome("");
+                            setResult(null);
+                            setExpanded(false);
+                        }}
+                        className="calcpage-btn danger"
+                    >
+                        🧹 נקה טופס
+                    </button>
+                </div>
 
                 {/* Unauthorized */}
                 {authError && (
@@ -114,10 +139,7 @@ export default function SelfEmployedCost() {
                         <AlertTriangle className="w-5 h-5 text-red-600" />
                         <span>
                             עליך להיות מחובר כדי לבצע חישוב.{" "}
-                            <button
-                                onClick={() => navigate("/SigninForm")}
-                                className="link-btn"
-                            >
+                            <button onClick={() => navigate("/SigninForm")} className="link-btn">
                                 התחבר כאן
                             </button>
                         </span>
@@ -129,26 +151,26 @@ export default function SelfEmployedCost() {
                     <div className="calcpage-result">
                         <h3>תוצאות החישוב</h3>
 
-                        {/* Summary */}
+                        {/* Summary cards */}
                         <div className="calcpage-summary-cards">
                             <div className="summary-card blue">
-                                <h4>תשלום חודשי</h4>
-                                <p>{result.national_insurance.monthly_prepayment.toLocaleString()} ₪</p>
+                                <h4>סה״כ חודשי</h4>
+                                <p>{fmt(result.summary.monthly_prepayment)} ₪</p>
                             </div>
                             <div className="summary-card red">
-                                <h4>תשלום שנתי</h4>
-                                <p>{result.national_insurance.yearly_total.toLocaleString()} ₪</p>
+                                <h4>סה״כ שנתי</h4>
+                                <p>{fmt(result.summary.yearly_total)} ₪</p>
                             </div>
                             <div className="summary-card green">
-                                <h4>נטו אחרי ביטוח לאומי</h4>
-                                <p>{result.summary.net_after_ni.toLocaleString()} ₪</p>
+                                <h4>נטו אחרי ניכויים</h4>
+                                <p>{fmt(result.summary.net_after_deductions)} ₪</p>
                             </div>
                         </div>
 
                         {/* Expand details */}
                         <button
                             className="expand-btn"
-                            onClick={() => setExpanded((prev) => !prev)}
+                            onClick={() => setExpanded((p) => !p)}
                         >
                             {expanded ? (
                                 <>
@@ -163,77 +185,60 @@ export default function SelfEmployedCost() {
 
                         {expanded && (
                             <div className="details-box">
-                                <h4>פירוט ביטוח לאומי</h4>
+                                <h4>קלט</h4>
                                 <ul>
-                                    <li>הגדרה: {result.national_insurance.definition}</li>
-                                    <li>חלק נמוך: {result.national_insurance.breakdown.low_rate_part} ₪</li>
-                                    <li>חלק גבוה: {result.national_insurance.breakdown.high_rate_part} ₪</li>
+                                    <li>הכנסה שנתית: {fmt(result.inputs.yearly_income)} ₪</li>
+                                    <li>הכנסה חייבת (70%): {fmt(result.inputs.taxable_income)} ₪</li>
+                                </ul>
+
+                                <h4>פירוט ניכויים</h4>
+                                <ul>
+                                    <li>ביטוח לאומי: {fmt(result.breakdown.national_insurance)} ₪</li>
+                                    <li>ביטוח בריאות: {fmt(result.breakdown.health_insurance)} ₪</li>
                                 </ul>
                             </div>
                         )}
                     </div>
                 )}
-
-                {/* Footer Buttons */}
-                <div className="calcpage-form-footer">
-                    <button onClick={() => navigate(-1)} className="calcpage-btn home">
-                        🔙 חזור
-                    </button>
-                    <button
-                        onClick={() => {
-                            setYearlyIncome("");
-                            setResult(null);
-                        }}
-                        className="calcpage-btn danger"
-                    >
-                        🧹 נקה טופס
-                    </button>
-                </div>
             </section>
 
             {/* Description */}
             <section className="calcpage-description">
-                <h2>מה זה מחשבון עצמאי - ביטוח לאומי?</h2>
+                <h2>מה זה מחשבון עצמאי זעיר?</h2>
                 <p>
-                    מחשבון עצמאי - ביטוח לאומי הוא כלי מתמחה לחישוב תשלומי ביטוח לאומי עבור עצמאים.
-                    המחשבון מחשב את התשלום החודשי והשנתי לביטוח לאומי על בסיס ההכנסה השנתית,
-                    ומציג את הנטו שנותר לאחר ניכוי ביטוח לאומי. זהו כלי חיוני לכל עצמאי המעוניין להבין את חבויותיו.
+                    מחשבון עצמאי זעיר הוא כלי מתמחה לחישוב תשלומי ביטוח לאומי ובריאות עבור עצמאים זעירים.
+                    המחשבון מחשב את התשלומים על בסיס ההכנסה החייבת (70% מהברוטו) ומציג תשלום חודשי,
+                    תשלום שנתי, ונטו לאחר הניכויים. זהו כלי חיוני לכל עצמאי זעיר המעוניין להבין את חבויותיו.
                 </p>
 
-                <h2>למי מתאים מחשבון עצמאי - ביטוח לאומי?</h2>
+                <h2>למי מתאים מחשבון עצמאי זעיר?</h2>
                 <ul>
-                    <li>עצמאים בתחילת הדרך המעוניינים להבין את חבויותיהם</li>
-                    <li>פרילנסרים בתחומים שונים</li>
-                    <li>עוסקים מורשים עם לקוחות קבועים</li>
-                    <li>עצמאים המעוניינים לתכנן את תשלומיהם השנתיים</li>
+                    <li>עוסקים פטורים בתחילת הדרך</li>
+                    <li>פרילנסרים עם הכנסה נמוכה עד בינונית</li>
+                    <li>עצמאים זעירים המעוניינים לתכנן את תשלומיהם</li>
+                    <li>יועצים המסייעים לעצמאים זעירים</li>
                 </ul>
 
-                <h2>למה כדאי להשתמש במחשבון עצמאי - ביטוח לאומי?</h2>
+                <h2>למה כדאי להשתמש במחשבון עצמאי זעיר?</h2>
                 <ul>
-                    <li>✔️ חישוב מדויק של תשלומי ביטוח לאומי</li>
+                    <li>✔️ חישוב מהיר ומדויק של תשלומי ביטוח לאומי ובריאות</li>
                     <li>✔️ תצוגה ברורה של תשלום חודשי ושנתי</li>
                     <li>✔️ עוזר בתכנון תזרים מזומנים</li>
-                    <li>✔️ מאפשר הבנה מלאה של חבויות העצמאי</li>
+                    <li>✔️ מאפשר הבנה מלאה של חבויות העצמאי הזעיר</li>
                 </ul>
 
-                <h2>מה תוכלו לגלות במחשבון עצמאי - ביטוח לאומי?</h2>
-                <ul>
-                    <li>מה גובה התשלום החודשי לביטוח לאומי</li>
-                    <li>מה ההפרש בין מדרגת תשלום נמוכה לגבוהה</li>
-                    <li>כמה נשאר נטו אחרי ניכוי ביטוח לאומי</li>
-                    <li>פירוט מלא של חישוב ביטוח לאומי</li>
-                </ul>
-
-                <h2>איך עובד מחשבון עצמאי - ביטוח לאומי?</h2>
+                <h2>איך עובד מחשבון עצמאי זעיר?</h2>
                 <p>
-                    המחשבון מקבל את ההכנסה השנתית שלכם ומחשב את תשלומי הביטוח הלאומי בהתאם למדרגות התשלום.
-                    הוא מציג את התשלום החודשי, השנתי, והנטו שנותר לאחר ניכוי ביטוח לאומי.
+                    המחשבון מקבל את ההכנסה השנתית הברוטו שלכם ומחשב את ההכנסה החייבת (70% מהברוטו).
+                    על בסיס זה הוא מחשב את תשלומי הביטוח הלאומי והבריאות, ומציג לכם את התשלום החודשי,
+                    השנתי, והנטו שנותר לאחר הניכויים.
                 </p>
 
                 <h2>כתב ויתור</h2>
                 <p className="disclaimer">
-                    מחשבון עצמאי - ביטוח לאומי נותן הערכה בלבד ואינו מהווה ייעוץ מס או תחליף לליווי מקצועי.
-                    לפני החלטות כלכליות חשובות, מומלץ להתייעץ עם רואה חשבון מוסמך לקבלת ייעוץ מותאם אישית.
+                    מחשבון עצמאי זעיר מספק אומדן בלבד ואינו מהווה ייעוץ מס או תחליף לליווי מקצועי.
+                    שיעורי הביטוח הלאומי והבריאות משתנים לפי חוק, והחישוב אינו כולל כל התנאים האישיים.
+                    מומלץ להתייעץ עם רואה חשבון או יועץ מס מוסמך לקבלת ייעוץ מותאם אישית.
                 </p>
             </section>
         </div>
