+++ src/components/Layout/Footer.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./Styles/Footer.css";
+import "./Footer.css";
 
 function Footer() {
   return (
