+++ src/pages/AboutUs.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../styles/AboutUs.css";
+import "./AboutUs.css";
 import logoImage from "../assets/logo.png";
 
 function AboutUs() {
