--- src/pages/HomePage2.jsx
@@ -1,29 +1,0 @@
-import React from "react";
-import HeroSection from "../components/HomePage/HeroSection";
-import ServicesSection from "../components/HomePage/ServicesSection";
-import WhySection from "../components/HomePage/WhySection";
-import CoursesSection from "../components/HomePage/CoursesSection";
-import SuccessSection from "../components/HomePage/SuccessSection";
-import JourneySection from "../components/HomePage/JourneySection";
-import FAQSection from "../components/HomePage/FAQSection";
-import TestimonialsSection from "../components/HomePage/TestimonialsSection";
-import ArticlesSection from "../components/HomePage/ArticlesSection";
-
-function HomePage2() {
-  return (
-    <div>
-      <HeroSection />
-      <ServicesSection />
-      <WhySection />
-      <CoursesSection />
-      <SuccessSection />
-      <JourneySection />
-      <FAQSection />
-      <TestimonialsSection />
-      <ArticlesSection />
-    </div>
-  );
-}
-
-export default HomePage2;
-
