+++ src/pages/UserPage.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import "../styles/UserPage.css";
+import "./UserPage.css";
 
 const UserPage = () => {
     const navigate = useNavigate();
