--- src/styles/Home3.css
@@ -1,109 +1,0 @@
-/* === Header Bar === */
-.navbar {
-    position: sticky;
-    top: 0;
-    z-index: 50;
-    background: #fff;
-    border-bottom: 1px solid #e5eaf5;
-    /* subtle border */
-}
-
-/* Layout */
-.navbar-container {
-    direction: rtl;
-    /* keep brand on the right */
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 10px 24px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
-
-/* === Brand === */
-.navbar-brand {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    text-decoration: none;
-}
-
-.brand-icon {
-    display: inline-grid;
-    place-items: center;
-    width: 36px;
-    height: 36px;
-    border-radius: 8px;
-    background: #2563eb;
-    /* blue */
-    color: #fff;
-    font-size: 18px;
-    box-shadow: 0 4px 10px rgba(37, 99, 235, 0.25);
-}
-
-.brand-text {
-    display: flex;
-    flex-direction: column;
-    line-height: 1.1;
-}
-
-.brand-text strong {
-    font-size: 15px;
-    font-weight: 700;
-    color: #1e293b;
-    /* slate-800 */
-}
-
-.brand-text small {
-    font-size: 12px;
-    color: #64748b;
-    /* slate-500 */
-}
-
-/* === Links === */
-.navbar-links {
-    display: flex;
-    gap: 20px;
-    align-items: center;
-}
-
-.navbar-link {
-    font-size: 14px;
-    font-weight: 500;
-    color: #475569;
-    /* slate-600 */
-    text-decoration: none;
-    padding: 6px 12px;
-    border-radius: 8px;
-    transition: 0.2s ease all;
-}
-
-.navbar-link:hover {
-    background: #f0f6ff;
-    color: #2563eb;
-}
-
-.navbar-link.active {
-    background: #2563eb;
-    color: #fff;
-    box-shadow: 0 4px 10px rgba(37, 99, 235, 0.25);
-}
-
-/* === Contact Button === */
-.btn-subscribe {
-    padding: 8px 16px;
-    background: linear-gradient(to right, #f59e0b, #d97706);
-    /* amber */
-    color: #fff;
-    border: none;
-    border-radius: 8px;
-    font-weight: 600;
-    text-decoration: none;
-    box-shadow: 0 4px 10px rgba(245, 158, 11, 0.25);
-    transition: 0.2s ease all;
-}
-
-.btn-subscribe:hover {
-    box-shadow: 0 6px 12px rgba(245, 158, 11, 0.35);
-    transform: translateY(-1px);
-}
